@@ -3,29 +3,31 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@date-io/date-fns": "^1.3.13",
-    "@hookform/resolvers": "^2.8.0",
-    "@material-ui/core": "^4.12.3",
-    "@material-ui/icons": "^4.11.2",
-    "@material-ui/lab": "^4.0.0-alpha.60",
-    "@material-ui/pickers": "^3.3.10",
-    "@material-ui/styles": "^4.11.4",
-    "@testing-library/jest-dom": "^5.14.1",
-    "@testing-library/react": "^11.2.7",
-    "@testing-library/user-event": "^12.8.3",
-    "axios": "^0.21.1",
-    "date-fns": "^2.23.0",
-    "mui-datatables": "^3.7.8",
-    "prettier": "^2.8.8",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-hook-form": "^7.14.1",
-    "react-number-format": "^4.7.3",
-    "react-router-dom": "^5.2.0",
-    "react-scripts": "4.0.3",
-    "recharts": "^2.1.8",
-    "web-vitals": "^1.1.2",
-    "yup": "^0.32.9"
+    "@date-io/date-fns": "^3.0.0",
+    "@emotion/react": "^11.11.4",
+    "@emotion/styled": "^11.11.5",
+    "@hookform/resolvers": "^3.3.4",
+    "@mui/icons-material": "^5.15.14",
+    "@mui/lab": "^5.0.0-alpha.169",
+    "@mui/material": "^5.15.14",
+    "@mui/styles": "^5.15.14",
+    "@mui/x-date-pickers": "^7.1.0",
+    "@testing-library/jest-dom": "^6.4.2",
+    "@testing-library/react": "^14.2.2",
+    "@testing-library/user-event": "^14.5.2",
+    "axios": "^1.6.8",
+    "date-fns": "^3.6.0",
+    "mui-datatables": "^4.3.0",
+    "prettier": "^3.1.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-hook-form": "^7.51.2",
+    "react-number-format": "^5.3.4",
+    "react-router-dom": "^6.22.3",
+    "react-scripts": "^5.0.1",
+    "recharts": "^2.12.3",
+    "web-vitals": "^3.5.2",
+    "yup": "^1.4.0"
   },
   "scripts": {
     "start": "react-scripts start",
@@ -50,5 +52,11 @@
       "last 1 firefox version",
       "last 1 safari version"
     ]
+  },
+  "devDependencies": {
+    "@typescript-eslint/parser": "^5.0.1",
+    "eslint": "^8.52.0",
+    "prettier-eslint": "^16.1.2",
+    "typescript": "^4.4.4"
   }
 }
