@@ -3,25 +3,24 @@ import { useState } from "react";
 import MUIDataTable from "mui-datatables";
 
 import DateFnsUtils from "@date-io/date-fns";
-import {
-  MuiPickersUtilsProvider,
-  KeyboardDatePicker,
-} from "@material-ui/pickers";
 
-import Button from "@material-ui/core/Button";
-import CircularProgress from "@material-ui/core/CircularProgress";
-import FormGroup from "@material-ui/core/FormGroup";
-import FormLabel from "@material-ui/core/FormLabel";
-import Dialog from "@material-ui/core/Dialog";
-import DialogActions from "@material-ui/core/DialogActions";
-import DialogContent from "@material-ui/core/DialogContent";
-import DialogTitle from "@material-ui/core/DialogTitle";
+import { LocalizationProvider } from '@mui/x-date-pickers/LocalizationProvider';
+import DatePicker from '@mui/lab/DatePicker';
 
-import AddIcon from "@material-ui/icons/Add";
-import DeleteIcon from "@material-ui/icons/Delete";
-import EditIcon from "@material-ui/icons/Edit";
-import IconButton from "@material-ui/core/IconButton";
-import Container from "@material-ui/core/Container";
+import Button from "@mui/material/Button";
+import CircularProgress from "@mui/material/CircularProgress";
+import FormGroup from "@mui/material/FormGroup";
+import FormLabel from "@mui/material/FormLabel";
+import Dialog from "@mui/material/Dialog";
+import DialogActions from "@mui/material/DialogActions";
+import DialogContent from "@mui/material/DialogContent";
+import DialogTitle from "@mui/material/DialogTitle";
+
+import AddIcon from "@mui/icons-material/Add";
+import DeleteIcon from "@mui/icons-material/Delete";
+import EditIcon from "@mui/icons-material/Edit";
+import IconButton from "@mui/material/IconButton";
+import Container from "@mui/material/Container";
 
 import { ExpenseForm } from "../../forms/ExpenseForm";
 import { FormFeedback } from "../FormFeedback";
@@ -33,8 +32,8 @@ import {
   ExpensesSourcesMapping,
 } from "../../consts.js";
 
-import PlusOneIcon from "@material-ui/icons/PlusOne";
-import Tooltip from "@material-ui/core/Tooltip";
+import PlusOneIcon from "@mui/icons-material/PlusOne";
+import Tooltip from "@mui/material/Tooltip";
 
 const ExpenseCreateEditDialog = ({ data, open, onClose, reloadTable }) => {
   return (
@@ -329,10 +328,10 @@ export const ExpensesTable = () => {
         filterOptions: {
           names: [],
           display: (filterList, onChange, index, column) => (
-            <MuiPickersUtilsProvider utils={DateFnsUtils}>
+            <LocalizationProvider utils={DateFnsUtils}>
               <FormLabel>Quando</FormLabel>
               <FormGroup row>
-                <KeyboardDatePicker
+                <DatePicker
                   disableToolbar
                   variant="inline"
                   format="dd/MM/yyyy"
@@ -348,7 +347,7 @@ export const ExpensesTable = () => {
                   }}
                   style={{ width: "48%", marginRight: "2%" }}
                 />
-                <KeyboardDatePicker
+                <DatePicker
                   disableToolbar
                   variant="inline"
                   format="dd/MM/yyyy"
@@ -365,7 +364,7 @@ export const ExpensesTable = () => {
                   style={{ width: "48%" }}
                 />
               </FormGroup>
-            </MuiPickersUtilsProvider>
+            </LocalizationProvider>
           ),
         },
       },
