@@ -1,5 +1,5 @@
-import Grid from "@material-ui/core/Grid";
-import Typography from "@material-ui/core/Typography";
+import Grid from "@mui/material/Grid";
+import Typography from "@mui/material/Typography";
 
 import { UserProfileForm } from "../../forms/UserProfileForm";
 import { ChangePasswordForm } from "../../forms/ChangePasswordForm";
