@@ -1,24 +1,21 @@
 import { useEffect, useState } from "react";
 
 import { useForm, Controller } from "react-hook-form";
-import NumberFormat from "react-number-format";
+import { NumericFormat } from "react-number-format";
 import * as yup from "yup";
 
 import DateFnsUtils from "@date-io/date-fns";
-import {
-  MuiPickersUtilsProvider,
-  KeyboardDatePicker,
-} from "@material-ui/pickers";
+import { LocalizationProvider } from '@mui/x-date-pickers/LocalizationProvider';
+import DatePicker from '@mui/lab/DatePicker';
+import Autocomplete from "@mui/lab/Autocomplete";
 
-import Autocomplete from "@material-ui/lab/Autocomplete";
-
-import Button from "@material-ui/core/Button";
-import CircularProgress from "@material-ui/core/CircularProgress";
-import DialogActions from "@material-ui/core/DialogActions";
-import FormControl from "@material-ui/core/FormControl";
-import FormGroup from "@material-ui/core/FormGroup";
-import FormHelperText from "@material-ui/core/FormHelperText";
-import TextField from "@material-ui/core/TextField";
+import Button from "@mui/material/Button";
+import CircularProgress from "@mui/material/CircularProgress";
+import DialogActions from "@mui/material/DialogActions";
+import FormControl from "@mui/material/FormControl";
+import FormGroup from "@mui/material/FormGroup";
+import FormHelperText from "@mui/material/FormHelperText";
+import TextField from "@mui/material/TextField";
 
 import { yupResolver } from "@hookform/resolvers/yup";
 
@@ -34,7 +31,7 @@ function NumberFormatCustom(props) {
   const { inputRef, onChange, ...other } = props;
 
   return (
-    <NumberFormat
+    <NumericFormat
       {...other}
       getInputRef={inputRef}
       onValueChange={(values) =>
@@ -48,7 +45,7 @@ function NumberFormatCustom(props) {
       decimalSeparator=","
       decimalScale={2}
       allowNegative={false}
-      isNumericString
+      valueIsNumericString
     />
   );
 }
@@ -339,7 +336,7 @@ export const PassiveIncomeForm = ({
               )}
             />
           </FormControl>
-          <MuiPickersUtilsProvider utils={DateFnsUtils}>
+          <LocalizationProvider utils={DateFnsUtils}>
             <Controller
               name="operation_date"
               control={control}
@@ -350,7 +347,7 @@ export const PassiveIncomeForm = ({
                   : new Date()
               }
               render={({ field: { onChange, value } }) => (
-                <KeyboardDatePicker
+                <DatePicker
                   onChange={onChange}
                   value={value}
                   label="Quando?"
@@ -368,7 +365,7 @@ export const PassiveIncomeForm = ({
                 />
               )}
             />
-          </MuiPickersUtilsProvider>
+          </LocalizationProvider>
         </FormGroup>
         <FormGroup row style={{ marginTop: "10px" }}>
           <Controller
